move shared fields to a different file
